requirements.txt
torch
torchvision
numpy
opencv-python
Pillow
plotly
matplotlib
skimage
albumentations
regex
pycocotools
pandas
zlib-state
pybase64